refactor: split inference request preparation
